Add script for playing with xml exported from IRis.
Turns out that the xml is malformed, which makes it hard to use from a
script. I think the main issue is use of unicode characters, so we might
be able to pre-transform the xml files by replacing those chars with
their encoded equivalents. I'm not sure why that's necessary if the xml
specifies that it's using UTF-8 encoding, but python's xml library
doesn't like any of the files I've tried.